+++ src/wetterdienst/provider/metno/frost/__init__.py
@@ -1,0 +1,3 @@
+from wetterdienst.provider.metno.frost.api import MetnoFrostRequest
+
+__all__ = ["MetnoFrostRequest"]